fastapi-ai-chat/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
chromadb==0.4.24
pypdf==5.1.0
openai==1.57.0
groq==0.9.0
google-generativeai==0.7.2
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
numpy==1.26.4  # pin <2.0 to avoid removed np.float_ used by chromadb
bcrypt==4.0.1  # known-good for passlib bcrypt backend

